Fixed Dungeon Clear logic handler treating dungeon clear logic as vanilla if it's randomization status was unknown.
Fixed the _ key in filter search considering items valid if they didn't have a randomized item.

Fixed the Rando setting importer not marking starting items as junk if the setting is enabled.

Updated the image for the royal wallet. The picture is pretty bad since it was a screenshot from in game, but at least it's the correct image.

Fixed the check mark beside the "Show location Item name" option in the logic editor being reverse.

Readded "GameClearDungeonEntrance" entries to the V9 and 10 logic dictionaries since they got removed for some reason.