Cleaned up the method used by the logic editor to differentiate between items with an identical name.
Added a function that will take in an item name and if there are multiple entries in logic with the same item name add the location name in parenthasis after the item name. Otherwise return the item name (or dictionary name if the item name is null).

This function is used when the logic eidtor writes an item to a list instead of actually changing the item name when initialized. This caused issues as that new name would propigate if the logic editor logic was applied to the actual tracker.