Fixed some code to allow for at least basic funtionality when running under WINE on Mac or Linux systems.
File paths to various recources are now always include the full system path using "AppDomain.CurrentDomain.BaseDirectory". It seems Wine does not correctly append the current working directory when given a file path without it. This fixes that and hopefully makes the code a bit more solid overall.

Json Logic files are now parsed with Newtonsoft instead of the method in the LogicFile class. It seems there is something in the Serializer options that WINE doesn't like. This serializer is still used when saving logic in the logic editor, so that currently doesn't work. This needs use that method until I'm confident I can recreate the formatting with NewtonSoft.

===================
KNOWN INCOMPATIBILITIES

-As described above, saving logic via the logic editor will crash the program

-Wine seems to be unable to make web requests to some websites including Github. Because of this the MMR presets will not work. This also means that importing a spoiler log as logic will not work if the logic used was casual or glitched. A standard logic file must be imported first and then the spoiler log on top of it.