Logic file JSON parsing operations now only use Newtonsoft instead of System.Text.Json. This should fix compatibility with WINE.
Added a refresh button to the main form (can also be triggered with F5). This will refresh all UI elements and perform a logic caluclation. This can be useful for UI bugs or logic bugs. Its also useful when running in WINE since form resize triggers don't work.

Fixed the form not properly resizing when it maximized and unmaximized.